refactor: tableGroup 패키지 분
