+++ src/test/java/kitchenpos/integration/TableIntegrationTest.java
@@ -6,9 +6,9 @@ import kitchenpos.order.OrderLineItem;
 import kitchenpos.order.OrderRepository;
 import kitchenpos.order.OrderStatus;
 import kitchenpos.table.OrderTable;
-import kitchenpos.table.TableGroup;
-import kitchenpos.table.TableGroupService;
 import kitchenpos.table.TableService;
+import kitchenpos.tablegroup.TableGroup;
+import kitchenpos.tablegroup.TableGroupService;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
